requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.32
bcrypt==4.0.1
pydub==0.25.1
Werkzeug==2.2.3
twilio>=7.0,<8.0
pyst2==0.5.1
